+++ Raaz/Core/Parse/Applicative.hs
@@ -11,6 +11,7 @@ module Raaz.Core.Parse.Applicative
        ) where
 
 import           Data.ByteString           (ByteString)
+import           Data.Proxy
 import           Data.Vector.Generic       (Vector, generateM)
 import           Foreign.Ptr               (castPtr)
 import           Foreign.Storable          (Storable, peek, peekElemOff)
@@ -57,8 +58,8 @@ unsafeRunParser :: Parser a -> Pointer -> IO a
 unsafeRunParser = runFieldM . twistFunctorValue
 
 -- | The primary purpose of this function is to satisfy type checkers.
-undefParse :: Parser a -> a
-undefParse _ = undefined
+parserToProxy   :: Parser a -> Proxy a
+parserToProxy _ = Proxy
 
 -- | Parses a value which is an instance of Storable. Beware that this
 -- parser expects that the value is stored in machine endian. Mostly
@@ -66,7 +67,7 @@ undefParse _ = undefined
 -- `Storable` instance.
 parseStorable :: Storable a => Parser a
 parseStorable = pa
-  where pa = makeParser (sizeOf $ undefParse pa) (peek . castPtr)
+  where pa = makeParser (sizeOf $ parserToProxy pa) (peek . castPtr)
 
 -- | Parse a crypto value. Endian safety is take into account
 -- here. This is what you would need when you parse packets from an
@@ -74,7 +75,7 @@ parseStorable = pa
 -- function in a complicated `EndianStore` instance.
 parse :: EndianStore a => Parser a
 parse = pa
-  where pa = makeParser (sizeOf $ undefParse pa) (load . castPtr)
+  where pa = makeParser (sizeOf $ parserToProxy pa) (load . castPtr)
 
 -- | Parses a strict bytestring of a given length.
 parseByteString :: LengthUnit l => l -> Parser ByteString
@@ -87,10 +88,10 @@ parseByteString l = makeParser l $ createFrom l
 unsafeParseStorableVector :: (Storable a, Vector v a) => Int -> Parser (v a)
 unsafeParseStorableVector n = pvec
   where pvec      = makeParser  width $ \ cptr -> generateM n (getA cptr)
-        width     = fromIntegral n * sizeOf (undefA pvec)
-        undefA    :: (Storable a, Vector v a)=> Parser (v a) -> a
-        undefA _  = undefined
+        width     = fromIntegral n * sizeOf (thisProxy pvec)
         getA      = peekElemOff . castPtr
+        thisProxy    :: Storable a => Parser (v a) -> Proxy a
+        thisProxy _ = Proxy
 
 -- | Similar to @parseVector@ but is expected to be slightly
 -- faster. It does not check whether the length parameter is
@@ -99,9 +100,9 @@ unsafeParseStorableVector n = pvec
 unsafeParseVector :: (EndianStore a, Vector v a) => Int -> Parser (v a)
 unsafeParseVector n = pvec
   where pvec     = makeParser  width $ \ cptr -> generateM n (loadFromIndex (castPtr cptr))
-        width    = fromIntegral n * sizeOf (undefA pvec)
-        undefA   :: (EndianStore a, Vector v a)=> Parser (v a) -> a
-        undefA _ = undefined
+        width    = fromIntegral n * sizeOf (thisProxy pvec)
+        thisProxy    :: Storable a => Parser (v a) -> Proxy a
+        thisProxy _ = Proxy
 
 -- | Similar to `parseVector` but parses according to the host
 -- endian. This function is essentially used to define storable